Ionisation product searches of a metabolite database

Methods for metabolite database ionisation product searches

Usage

calcAdducts(db, id, adduct_rules_table = adduct_rules())

# S4 method for MetaboliteDatabase
calcAdducts(db, id, adduct_rules_table = adduct_rules())

PIPsearch(
  db,
  mz,
  adduct,
  ppm = 6,
  isotope = NA,
  adduct_rules_table = adduct_rules(),
  isotope_rules_table = isotope_rules()
)

# S4 method for MetaboliteDatabase
PIPsearch(
  db,
  mz,
  adduct,
  ppm = 6,
  isotope = NA,
  adduct_rules_table = adduct_rules(),
  isotope_rules_table = isotope_rules()
)

Arguments

db

S4 object of class MetaboliteDatabase

id

Database entry ID

adduct_rules_table

table containing adduct formation rules. Defaults to adduct_rules().

mz

mass to charge ratio to search

adduct

the search adduct as available in mzAnnotation::adduct_names()

ppm

the parts per million error search threshold

isotope

the search isotope as available in mzAnnotation::isotope_names()

isotope_rules_table

table containing isotope rules. Defaults to isotope_rules().

Value

A tibble containing the relevant product search results depending on the method used.

Details

PIPsearch - perform a putative ionisation product search on a metabolite database

calcAdducts - calculate adduct mass to charge ratios for a metabolite database entry

Examples

metabolite_database <- metaboliteDB(amino_acids)

## Perform a putative ionisation product search
PIPsearch(
  metabolite_database,
  mz = 133.03358,
  adduct = '[M-H]1-',
  isotope = '13C')
#> # A tibble: 1 × 12
#>      ID NAME   InChI InChI…¹ SMILES MF    Accur…² Isotope Adduct Measu…³ Theor…⁴
#>   <int> <chr>  <chr> <chr>   <chr>  <chr>   <dbl> <chr>   <chr>    <dbl>   <dbl>
#> 1     4 L-Asp… InCh… CKLJMW… C([C@… C4H7…    133. 13C     [M-H]…    133.    133.
#> # … with 1 more variable: `PPM Error` <dbl>, and abbreviated variable names
#> #   ¹​InChIKey, ²​Accurate_Mass, ³​`Measured m/z`, ⁴​`Theoretical m/z`

## Calculate adduct m/z for a database entry
calcAdducts(metabolite_database,1)
#> # A tibble: 58 × 4
#>    Adduct         `m/z` MF         Possible
#>    <chr>          <dbl> <chr>      <lgl>   
#>  1 [M+3Na]3+       52.7 C3H7NNa3O2 TRUE    
#>  2 [M+H+2Na]3+     45.3 C3H8NNa2O2 TRUE    
#>  3 [M+2H+Na]3+     38.0 C3H9NNaO2  TRUE    
#>  4 [M+3H]3+        30.7 C3H10NO2   TRUE    
#>  5 [2M+3H2O+2H]2+ 117.  C6H22N2O7  TRUE    
#>  6 [M+3ACN+2H]2+  107.  C9H18N4O2  TRUE    
#>  7 [M+2ACN+2H]2+   86.6 C7H15N3O2  TRUE    
#>  8 [M+2Na]2+       67.5 C3H7NNa2O2 TRUE    
#>  9 [M+ACN+2H]2+    66.0 C5H12N2O2  TRUE    
#> 10 [M+H+K]2+       64.5 C3H8KNO2   TRUE    
#> # … with 48 more rows